# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wikis API",
    "description": "AI-powered repository documentation generator",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ask": {
      "post": {
        "summary": "Ask",
        "operationId": "ask_api_v1_ask_post",
        "parameters": [
          {
            "name": "accept",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "application/json",
              "title": "Accept"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research": {
      "post": {
        "summary": "Research",
        "operationId": "research_api_v1_research_post",
        "parameters": [
          {
            "name": "accept",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "application/json",
              "title": "Accept"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wikis": {
      "get": {
        "summary": "List Wikis",
        "operationId": "list_wikis_api_v1_wikis_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikiListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/wikis/{wiki_id}": {
      "delete": {
        "summary": "Delete Wiki",
        "operationId": "delete_wiki_api_v1_wikis__wiki_id__delete",
        "parameters": [
          {
            "name": "wiki_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wiki Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWikiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AskRequest": {
        "description": "Request to ask a question about a wiki.",
        "properties": {
          "wiki_id": {
            "title": "Wiki Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "chat_history": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "Chat History",
            "type": "array"
          },
          "k": {
            "default": 15,
            "title": "K",
            "type": "integer"
          }
        },
        "required": [
          "wiki_id",
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResponse": {
        "description": "Response to an ask question.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "answer"
        ],
        "title": "AskResponse",
        "type": "object"
      },
      "ChatMessage": {
        "description": "A single chat message.",
        "properties": {
          "role": {
            "title": "Role",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
      "DeleteWikiResponse": {
        "description": "Response after deleting a wiki.",
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "wiki_id": {
            "title": "Wiki Id",
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "wiki_id"
        ],
        "title": "DeleteWikiResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResearchRequest": {
        "description": "Request to perform deep research on a wiki.",
        "properties": {
          "wiki_id": {
            "title": "Wiki Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "research_type": {
            "default": "general",
            "title": "Research Type",
            "type": "string"
          },
          "enable_subagents": {
            "default": true,
            "title": "Enable Subagents",
            "type": "boolean"
          }
        },
        "required": [
          "wiki_id",
          "question"
        ],
        "title": "ResearchRequest",
        "type": "object"
      },
      "ResearchResponse": {
        "description": "Response from deep research.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "title": "Sources",
            "type": "array"
          },
          "research_steps": {
            "items": {
              "type": "string"
            },
            "title": "Research Steps",
            "type": "array"
          }
        },
        "required": [
          "answer"
        ],
        "title": "ResearchResponse",
        "type": "object"
      },
      "SourceReference": {
        "description": "A reference to a source code location.",
        "properties": {
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "line_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Line Start"
          },
          "line_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Line End"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Snippet"
          }
        },
        "required": [
          "file_path"
        ],
        "title": "SourceReference",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WikiListResponse": {
        "description": "Response listing all wikis.",
        "properties": {
          "wikis": {
            "items": {
              "$ref": "#/components/schemas/WikiSummary"
            },
            "title": "Wikis",
            "type": "array"
          }
        },
        "title": "WikiListResponse",
        "type": "object"
      },
      "WikiSummary": {
        "description": "Summary of a generated wiki.",
        "properties": {
          "wiki_id": {
            "title": "Wiki Id",
            "type": "string"
          },
          "repo_url": {
            "title": "Repo Url",
            "type": "string"
          },
          "branch": {
            "title": "Branch",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "page_count": {
            "title": "Page Count",
            "type": "integer"
          },
          "status": {
            "default": "complete",
            "title": "Status",
            "type": "string"
          },
          "progress": {
            "default": 1.0,
            "title": "Progress",
            "type": "number"
          },
          "invocation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Invocation Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Error"
          }
        },
        "required": [
          "wiki_id",
          "repo_url",
          "branch",
          "title",
          "created_at",
          "page_count"
        ],
        "title": "WikiSummary",
        "type": "object"
      },
...
    }
  }
}